feat: add player_boot and rng_boot; refactor game.js to use createInitialPlayer and initRngRuntime
