+++ ROguelike 3_10_2025 filut jarjestyksessa luultavasti vikaa/core/engine/player_boot.js
@@ -1,0 +1,5 @@
+export function createInitialPlayer() {
+  return ((typeof window !== "undefined" && window.Player && typeof window.Player.createInitial === "function")
+    ? window.Player.createInitial()
+    : { x: 0, y: 0, hp: 20, maxHp: 40, inventory: [], atk: 1, xp: 0, level: 1, xpNext: 20, equipment: { left: null, right: null, head: null, torso: null, legs: null, hands: null } });
+}
